fix: organização de arquivos para webpack
