--- src/pages/home/home.css
@@ -1,4 +1,0 @@
-body {
-    color: rgb(28, 117, 111);
-    background-color: rgb(154, 154, 234);
-}
