--- src/pages/home/home.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-br">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Minha Página Inicial</title>
-</head>
-<body>
-    <h1>Home</h1>
-    <p>Bem-vindo à nossa página "Home". Aqui você pode encontrar um pouco de cada seção do nosso projeto.</p>
-</body>
-</html>
