--- src/pages/home/home.js
@@ -1,4 +1,0 @@
-// A importação de arquivos css são feitos dessa forma por causa do webpack
-import './home.css';
-
-alert('Aqui está a nossa página inicial!')
