+++ src/pages/login/index.html
@@ -1,15 +1,11 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>PS LAWD - Login</title>
 
-    <!--STYLE LOCAL-->
-    <link rel="stylesheet" type="text/css" href="login.css" media ="all">
-    
-    <!--STYLE GLOBAL-->
-    <link rel="stylesheet" type="text/css" href="../index.css" media ="all">
 
     <!--GOOGLE FONTS = SPACE GROTESK LIGHT 300-->
     <link rel="preconnect" href="https://fonts.googleapis.com">
@@ -19,14 +15,13 @@
     <!--ICONE DA ABA-->
     <!--<link rel="icon" type="image/png" href="../imagens/logo.png">-->
 
-    <!--SCRIPT GLOBAL-->
-    <script src="../index.js" defer></script>
 </head>
+
 <body>
     <header>
         <div class="icons">
-            <img src="../images/logo-lawd.png" class="img-icon"></img>
-            <img src="../images/lawd.png" class="img-lawd"></img>
+            <img src="../../assets/images/logo-lawd.png" class="img-icon"></img>
+            <img src="../../assets/images/lawd.png" class="img-lawd"></img>
         </div>
         <button>Cadastre-se</button>
     </header>
@@ -39,18 +34,19 @@
                 <label for="password">Senha</label>
                 <div class="password-container">
                     <input type="password" id="password" name="password" required>
-                    <img src="../images/visibility_off.png" class="password-visibility" id="visibility" onclick="visibilityPassword()">
+                    <img src="../../assets/images/visibility_off.png" class="password-visibility" id="visibility"
+                        onclick="visibilityPassword()">
                 </div>
                 <a href="##" class="password">Esqueceu a senha?</a>
             </form>
             <p>Ou entre com:</p>
             <div class="loginScreen-othersWays">
                 <div class="loginLink">
-                    <img src="../images/google.png" class="icon-login"></img>
+                    <img src="../../assets/images/google.png" class="icon-login"></img>
                     <p>Google</p>
                 </div>
                 <div class="loginLink">
-                    <img src="../images/github.png" class="icon-login"></img>
+                    <img src="../../assets/images/github.png" class="icon-login"></img>
                     <p>Github</p>
                 </div>
             </div>
@@ -61,4 +57,5 @@
 
     </footer>
 </body>
+
 </html>
