+++ src/pages/profile-evaluator/interview/index.css
@@ -1,0 +1,308 @@
+:root {
+    /*tons de roxos:*/
+    --purple: #4623A7;
+    --purple-dark: #3E2A78;
+    --purple-medium: #7046E9;
+    --purple-light: #BBA6F8;
+    --purple-white: #D7C9FF;
+    --purple-menu: #4A387B;
+    --purple-profile: #5b4a8c;
+    --purple-p: #8170B4;
+    --purple-result: #8F81B7;
+
+    /*tons de input:*/
+    --gray: #656565;
+    --background-input: #eff2ff73;
+
+    /*outros:*/
+    --gray-light: #CDCDCD;
+    --background-screens: #f9f8ff;
+    --gray-background: #EEEBF8;
+    --option: #eeecf6;
+
+}
+
+body {
+    background-image: url(images/background.svg);
+    background-size: cover;
+    background-repeat: no-repeat;
+    background-position: center;
+    background-attachment: fixed;
+    margin: 0;
+    padding: 0;
+    height: 100vh;
+    font-family: Space Grotesk;
+}
+
+header {
+    height: 20%;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 0em 1em;
+    padding: 0;
+}
+
+button {
+    font-family: Space Grotesk;
+    padding: 0.5em 1em;
+    background-color: var(--purple-light);
+    border: var(--purple-medium) solid 2px;
+    border-radius: 5em;
+    font-weight: 500;
+    color: var(--purple);
+    transition: background-color 0.8s, transform 0.7s;
+}
+
+button:hover {
+    background-color: var(--purple);
+    color: var(--purple-white);
+}
+
+button:active {
+    transform: scale(0.95);
+    background-color: var(--purple-white);
+    color: var(--purple);
+    border: var(--purple) solid 1px;
+}
+
+main {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+}
+
+main h1 {
+    color: #fff;
+    font-weight: 400;
+}
+
+form {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+}
+
+form label {
+    margin-top: 1.5em;
+    margin-bottom: 0.5em;
+    font-weight: bold;
+    color: var(--purple);
+    grid-column: span 1;
+}
+
+form input,
+form select {
+    border: solid var(--purple-white) 2px;
+    background-color: var(--background-input);
+    color: var(--gray);
+    padding: 0.2em 1em;
+    border-radius: 10px;
+    font-family: Space Grotesk;
+    grid-column: span 1;
+}
+
+form input:focus,
+.password-container:focus-within,
+form select:focus {
+    border-color: var(--purple);
+    outline: none;
+}
+
+p {
+    color: var(--gray);
+    font-weight: 300;
+    width: 100%;
+}
+
+footer {
+    height: 20%;
+}
+
+.lawd-background {
+    position: fixed;
+    bottom: 0;
+    right: 0;
+    width: clamp(200px, 55vw, 550px);
+    ;
+    z-index: -1;
+}
+
+.menu {
+    background: var(--purple-menu);
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    width: clamp(47em, 95%, 80em);
+    padding: 0.3em 0.7em;
+    border-radius: 30px;
+}
+
+.navbar {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-around;
+    align-items: center;
+    width: clamp(15em, 60%, 35em);
+    margin: 0;
+}
+
+.navbar li {
+    list-style-type: none;
+    color: #fff;
+    font-size: 0.7em;
+    width: 8em;
+    height: 2em;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    cursor: pointer;
+}
+
+.navbar li:not(.page):hover {
+    outline: solid #fff 1px;
+    border-radius: 10px;
+    box-sizing: border-box;
+}
+
+.page {
+    border: solid #fff 1px;
+    border-radius: 10px;
+    box-sizing: border-box;
+}
+
+.profile {
+    border: solid #fff 1px;
+    background: var(--purple-profile);
+    width: 8em;
+    padding: 0.2em 0.1em;
+    border-radius: 30px;
+    display: flex;
+    flex-direction: row;
+    gap: 0.5em;
+    justify-content: center;
+    align-items: center;
+    cursor: pointer;
+}
+
+.icon-user {
+    width: 30px;
+    height: 30px;
+    border-radius: 50%;
+}
+
+.name-user {
+    font-size: 0.8em;
+    display: inline-block;
+    max-width: 6em;
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    color: #fff;
+}
+
+.ps {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+    border: solid #fff 1px;
+    padding: 0 1em;
+    box-sizing: border-box;
+    border-radius: 25px;
+}
+
+.ps p {
+    font-size: 0.8em;
+    color: #fff;
+}
+
+.back {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    color: #fff;
+    text-decoration: none;
+    font-size: 0.8em;
+    gap: 3px;
+}
+
+.img-arrow {
+    width: 1.5em;
+    height: auto;
+}
+
+
+.icons {
+    display: flex;
+    align-items: center;
+    gap: 0.5em;
+    width: 8em;
+}
+
+.img-lawd {
+    width: 4em;
+    height: auto;
+}
+
+.img-icon {
+    width: 2em;
+    height: auto;
+}
+
+.password {
+    font-size: 0.7em;
+    text-align: right;
+    font-weight: 400;
+    text-decoration: underline;
+    color: var(--purple);
+}
+
+.password-container {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    background-color: var(--background-input);
+    border: solid var(--purple-white) 2px;
+    border-radius: 10px;
+}
+
+.password-container input {
+    width: 100%;
+    border: none;
+}
+
+.password-visibility {
+    width: 20px;
+    padding-right: 10px;
+    cursor: pointer;
+}
+
+
+@media (min-width: 1440px) {
+    .menu {
+        padding: 0.7em 1em;
+    }
+
+    .navbar li {
+        border-radius: 40px;
+        width: 8em;
+        height: 2.4em;
+        font-size: 0.8em;
+        padding: 0;
+    }
+
+    .navbar {
+        padding: 0;
+    }
+
+    .profile,
+    .icons {
+        width: 10em;
+    }
+
+    .name-user {
+        max-width: 8em;
+    }
+}
