+++ src/pages/profile-evaluator/interview/interview.html
@@ -7,7 +7,7 @@
     <title>PS LAWD - Entrevista</title>
 
     <!--STYLE GLOBAL-->
-    <link rel="stylesheet" type="text/css" href="../../../pages/index.css" media="all">
+    <link rel="stylesheet" type="text/css" href="index.css" media="all">
 
     <!--STYLE LOCAL-->
     <link rel="stylesheet" type="text/css" href="interview.css" media="all">
